tools/ai-sync/pyproject.toml
[project]
name = "ai-sync"
version = "0.1.0"
description = "Cross-runtime AI workflow sync tooling for agent-skills"
requires-python = ">=3.12"
dependencies = [
  "Jinja2>=3.1,<4",
  "PyYAML>=6.0,<7",
]
